updated code

+++ src/components/CurrenciesLayer.jsx
@@ -56,7 +56,7 @@ const CurrenciesLayer = () => {
     if (!dateString || isNaN(new Date(dateString).getTime())) return "";
     const date = new Date(dateString);
     const day = date.getDate();
-    const month = date.toLocaleString("en-GB", { month: "short" });
+    const month = date.toLocaleString("en-GB", { month: "long" });
     const year = date.getFullYear();
     const suffix = (day % 10 === 1 && day !== 11) ? "st" :
                    (day % 10 === 2 && day !== 12) ? "nd" :
