+++ src/components/StorageFacilityDetailsLayer.jsx
@@ -9,7 +9,7 @@ const formatDate = (dateString) => {
   if (!dateString || isNaN(new Date(dateString).getTime())) return "";
   const date = new Date(dateString);
   const day = date.getDate();
-  const month = date.toLocaleString("en-GB", { month: "short" });
+  const month = date.toLocaleString("en-GB", { month: "long" });
   const year = date.getFullYear();
   const suffix =
     day % 10 === 1 && day !== 11
